Rewrote the testsuite
The old test system based on an additional app, without it no testing was possible.
I didn't really liked this system, also I don't thing that Doctests should be placed
separate in a tests.py file.
To fix this issues, the doctests where replaced with unittest, also the testapp was
removed and the requiered test-model was added to the tests.py file. Not, every project
that uses metadata can run tests without another special testapp in there INSTALLED_APP
list.

Beside this, a new testapp was integrated, that allows a start of the tests directly
from within the metadata project, without any external additional project.